requirements.txt
# conda create -n rm python=3.11
# conda activate rm
# For Win/Linux: pip install torch==2.6.0+cu124 torchvision==0.21.0 torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu124
# For MAC: pip install torch==2.6.0 torchvision==0.21.0 torchaudio==2.6.0 --extra-index-url https://download.pytorch.org/whl/nightly/cpu
# pip install ../pkgs/flash_attn-2.7.4.post1+cu12torch2.6cxx11abiFALSE-cp311-cp311-linux_x86_64.whl
# pip install ../pkgs/flashinfer_python-0.2.2.post1+cu124torch2.6-cp38-abi3-linux_x86_64.whl
# pip install -r requirements.txt

accelerate
cleanlab[datalab]
codetiming
datasets<4.0.0
dill
einops
fastapi
huggingface_hub>=0.20.0
hydra-core
ipykernel
latex2sympy2_extended
# liger-kernel
math_verify
matplotlib
numpy<2.0.0
openai
packaging>=20.0
pandas
peft
pot
pre-commit
pyarrow>=19.0.0
pybind11
pylatexenc
pyyaml
ray[default]==2.47.1
safetensors
scikit-learn
scipy
seaborn
tensorboard
tensordict>=0.8.0,<=0.10.0,!=0.9.0
torchdata
transformers==4.55.4
uvicorn
vllm==0.8.5.post1
wandb